requirements.txt
gradio
mysql-connector-python
plotly